# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RDDecisionCard": {
        "properties": {
          "confidence": {
            "default": 0.6,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "next_experiment_zh": {
            "title": "Next Experiment Zh",
            "type": "string"
          },
          "rationale_zh": {
            "title": "Rationale Zh",
            "type": "string"
          },
          "recommendation_zh": {
            "title": "Recommendation Zh",
            "type": "string"
          },
          "risks_zh": {
            "items": {
              "type": "string"
            },
            "title": "Risks Zh",
            "type": "array"
          },
          "title_zh": {
            "title": "Title Zh",
            "type": "string"
          }
        },
        "required": [
          "title_zh",
          "recommendation_zh",
          "rationale_zh",
          "next_experiment_zh"
        ],
        "title": "RDDecisionCard",
        "type": "object"
      },
...
      "ScientificAnalysisRequest": {
        "properties": {
          "language_mode": {
            "$ref": "#/components/schemas/ScientificLanguageMode",
            "default": "zh_bilingual"
          },
          "objective_zh": {
            "default": "",
            "maxLength": 800,
            "title": "Objective Zh",
            "type": "string"
          },
          "session_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Session Ids",
            "type": "array"
          }
        },
        "required": [
          "session_ids"
        ],
        "title": "ScientificAnalysisRequest",
        "type": "object"
      },
...
      "ScientificClaim": {
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "claim_type": {
            "title": "Claim Type",
            "type": "string"
          },
          "condition_ids": {
            "items": {
              "type": "string"
            },
            "title": "Condition Ids",
            "type": "array"
          },
          "confidence": {
            "default": 0.7,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "evidence_quote": {
            "default": "",
            "title": "Evidence Quote",
            "type": "string"
          },
          "experiment_ids": {
            "items": {
              "type": "string"
            },
            "title": "Experiment Ids",
            "type": "array"
          },
          "metric_result_ids": {
            "items": {
              "type": "string"
            },
            "title": "Metric Result Ids",
            "type": "array"
          },
          "modality": {
            "default": "claimed",
            "enum": [
              "observed",
              "claimed",
              "hypothesized",
              "limited"
            ],
            "title": "Modality",
            "type": "string"
          },
          "object": {
            "default": "",
            "title": "Object",
            "type": "string"
          },
          "polarity": {
            "default": "neutral",
            "enum": [
              "positive",
              "negative",
              "neutral",
              "mixed"
            ],
            "title": "Polarity",
            "type": "string"
          },
          "predicate_zh": {
            "default": "",
            "title": "Predicate Zh",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "statement_original": {
            "default": "",
            "title": "Statement Original",
            "type": "string"
          },
          "statement_zh": {
            "title": "Statement Zh",
            "type": "string"
          },
          "subject": {
            "default": "",
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "claim_type",
          "statement_zh"
        ],
        "title": "ScientificClaim",
        "type": "object"
      },
      "ScientificCondition": {
        "properties": {
          "condition_id": {
            "title": "Condition Id",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "unit": {
            "default": "",
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "default": "",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "name"
        ],
        "title": "ScientificCondition",
        "type": "object"
      },
      "ScientificDocument": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "formulas": {
            "items": {
              "$ref": "#/components/schemas/ScientificFormula"
            },
            "title": "Formulas",
            "type": "array"
          },
          "page_dimensions": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Page Dimensions",
            "type": "object"
          },
          "reference_ids": {
            "items": {
              "type": "string"
            },
            "title": "Reference Ids",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ScientificSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "source_format": {
            "enum": [
              "jats",
              "tei",
              "grobid_tei"
            ],
            "title": "Source Format",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/ScientificTable"
            },
            "title": "Tables",
            "type": "array"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "source_format"
        ],
        "title": "ScientificDocument",
        "type": "object"
      },
      "ScientificEntity": {
        "properties": {
          "canonical_name": {
            "default": "",
            "title": "Canonical Name",
            "type": "string"
          },
          "description_zh": {
            "default": "",
            "title": "Description Zh",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "$ref": "#/components/schemas/ScientificEntityType"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "name_en": {
            "default": "",
            "title": "Name En",
            "type": "string"
          },
          "name_zh": {
            "title": "Name Zh",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "entity_type",
          "name_zh"
        ],
        "title": "ScientificEntity",
        "type": "object"
      },
      "ScientificEntityType": {
        "enum": [
          "research_problem",
          "method",
          "model",
          "dataset",
          "metric",
          "material",
          "parameter",
          "result",
          "limitation"
        ],
        "title": "ScientificEntityType",
        "type": "string"
      },
...
      "ScientificFormula": {
        "properties": {
          "formula_id": {
            "title": "Formula Id",
            "type": "string"
          },
          "locator": {
            "$ref": "#/components/schemas/ScientificLocator"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "formula_id",
          "text",
          "locator"
        ],
        "title": "ScientificFormula",
        "type": "object"
      },
...
      "ScientificLanguageMode": {
        "enum": [
          "zh",
          "zh_bilingual"
        ],
        "title": "ScientificLanguageMode",
        "type": "string"
      },
...
      "ScientificParseRequest": {
        "properties": {
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ScientificParseRequest",
        "type": "object"
      },
...
      "ScientificReport": {
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ScientificClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ScientificCondition"
            },
            "title": "Conditions",
            "type": "array"
          },
          "decision_cards": {
            "items": {
              "$ref": "#/components/schemas/RDDecisionCard"
            },
            "title": "Decision Cards",
            "type": "array"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/ScientificEntity"
            },
            "title": "Entities",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ScientificEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_matrix": {
            "items": {
              "$ref": "#/components/schemas/ScientificEvidenceMatrixRow"
            },
            "title": "Evidence Matrix",
            "type": "array"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ScientificExperiment"
            },
            "title": "Experiments",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "insights": {
            "items": {
              "$ref": "#/components/schemas/ScientificInsight"
            },
            "title": "Insights",
            "type": "array"
          },
          "language_mode": {
            "$ref": "#/components/schemas/ScientificLanguageMode",
            "default": "zh_bilingual"
          },
          "nary_relations": {
            "items": {
              "$ref": "#/components/schemas/ScientificNaryRelation"
            },
            "title": "Nary Relations",
            "type": "array"
          },
          "objective_zh": {
            "default": "",
            "title": "Objective Zh",
            "type": "string"
          },
          "papers": {
            "items": {
              "$ref": "#/components/schemas/ScientificPaperProfile"
            },
            "title": "Papers",
            "type": "array"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/ScientificRelation"
            },
            "title": "Relations",
            "type": "array"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "session_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Session Ids",
            "type": "array"
          },
          "title_zh": {
            "title": "Title Zh",
            "type": "string"
          }
        },
        "required": [
          "title_zh",
          "session_ids"
        ],
        "title": "ScientificReport",
        "type": "object"
      },
      "ScientificSection": {
        "properties": {
          "child_section_ids": {
            "items": {
              "type": "string"
            },
            "title": "Child Section Ids",
            "type": "array"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "title": "Path",
            "type": "array"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/ScientificSentence"
            },
            "title": "Sentences",
            "type": "array"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "section_id"
        ],
        "title": "ScientificSection",
        "type": "object"
      },
...
      "ScientificTable": {
        "properties": {
          "caption": {
            "default": "",
            "title": "Caption",
            "type": "string"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/ScientificTableCell"
            },
            "title": "Cells",
            "type": "array"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "locator": {
            "$ref": "#/components/schemas/ScientificLocator"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          }
        },
        "required": [
          "table_id"
        ],
        "title": "ScientificTable",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Corpus2Node API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scientific/documents/{session_id}": {
      "get": {
        "operationId": "list_scientific_documents_api_scientific_documents__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScientificDocument"
                  },
                  "title": "Response List Scientific Documents Api Scientific Documents  Session Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scientific Documents",
        "tags": [
          "scientific"
        ]
      }
    },
    "/api/scientific/parse": {
      "post": {
        "operationId": "parse_scientific_sources_api_scientific_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScientificParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScientificDocument"
                  },
                  "title": "Response Parse Scientific Sources Api Scientific Parse Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Scientific Sources",
        "tags": [
          "scientific"
        ]
      }
    },
    "/api/scientific/run": {
      "post": {
        "operationId": "run_scientific_route_api_scientific_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScientificAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScientificReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scientific Route",
        "tags": [
          "scientific"
        ]
      }
    },
    "/api/scientific/run/stream": {
      "post": {
        "operationId": "run_scientific_stream_api_scientific_run_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScientificAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scientific Stream",
        "tags": [
          "scientific"
        ]
      }
    },
    "/api/scientific/{report_id}": {
      "delete": {
        "operationId": "delete_scientific_report_api_scientific__report_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Scientific Report Api Scientific  Report Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scientific Report",
        "tags": [
          "scientific"
        ]
      },
      "get": {
        "operationId": "get_scientific_report_api_scientific__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScientificReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scientific Report",
        "tags": [
          "scientific"
        ]
      }
    },
...
    "/scientific/documents/{session_id}": {
      "get": {
        "operationId": "list_scientific_documents_scientific_documents__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScientificDocument"
                  },
                  "title": "Response List Scientific Documents Scientific Documents  Session Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scientific Documents",
        "tags": [
          "scientific"
        ]
      }
    },
    "/scientific/parse": {
      "post": {
        "operationId": "parse_scientific_sources_scientific_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScientificParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScientificDocument"
                  },
                  "title": "Response Parse Scientific Sources Scientific Parse Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Scientific Sources",
        "tags": [
          "scientific"
        ]
      }
    },
    "/scientific/run": {
      "post": {
        "operationId": "run_scientific_route_scientific_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScientificAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScientificReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scientific Route",
        "tags": [
          "scientific"
        ]
      }
    },
    "/scientific/run/stream": {
      "post": {
        "operationId": "run_scientific_stream_scientific_run_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScientificAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scientific Stream",
        "tags": [
          "scientific"
        ]
      }
    },
    "/scientific/{report_id}": {
      "delete": {
        "operationId": "delete_scientific_report_scientific__report_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Scientific Report Scientific  Report Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scientific Report",
        "tags": [
          "scientific"
        ]
      },
      "get": {
        "operationId": "get_scientific_report_scientific__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScientificReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scientific Report",
        "tags": [
          "scientific"
        ]
      }
    },
...
  }
}